Fix weird spell check syntax highlighting issues
Syntax highlighting for spell check does not adapt to the fg color of
the text.  So red background on cyan is hard to read.
And for spellCap (capitalize first word of sentence) the highlight bg box is
the SAME color as the word, so you can't see anything!

This removes the background and instead uses underline for first word
cap and forces fg color for spellBad that works with the red background.