feat: [sway] create 'swaymode'
Move all key bindings under a custom mode so that I no longer have to
hold down mod4/super (in addition to shift in some cases).
Now I can just hit super and enter the 'swaymode'.
This means I only need one super key and it can move to the top of the
keyboard were it is more comfortable to hit, since I won't need to combo
it with shift anymore.